requirements.txt
appium-python-client
pytest
pytest-html
